requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
numpy==1.26.4
PyMuPDF==1.24.6
openai>=1.30.0
tenacity==8.2.3
python-dotenv==1.0.1
SQLAlchemy>=2.0.29
psycopg[binary]>=3.1.18
pgvector>=0.2.5
alembic>=1.13.1
httpx==0.27.2

# Scheduler for background ingestion jobs
APScheduler==3.10.4

# HTML parsing for ingestion
beautifulsoup4>=4.12.3
lxml>=5.3.0
